Rename and restructure project to be easier to navigate from file paths.
